refactor toast component + remove unnecessary code
