+++ client/packages/lowcoder/src/comps/hooks/toastComp/index.tsx
@@ -1,0 +1,2 @@
+export { ToastComp } from "./toastComp";
+export type { ToastType } from "./toastConstants";
